Lavaplayer integration!
I started working on audio and have both the trackScheduler and AudioPlayerSendHandler imported and setup, all its missing is the actual methods for parsing a track and passing it to play in a server. Sadly that's the hardest part for me to figure out so for now we'll have to deal with what we have.